JIRA-JRTC-3017: [DPB] It shouldn't remove dependencies of configuration when use breakout command to change port speed
When the execution of a DPB command only changes the speed of ports but
will not change the number of ports(e.g. 4x10 to 4x25), it is not
required to remove port and create port again. This commit is to enhance
DPB to set port speed to Config_DB when the number of interfaces aren't
changed after executing port breakout command.