JRTC-3155: [VLAN][t0] VLAN member can not be remove from Vlan1000 by CLI command
Root Cause:
In CLI, it use the field 'members' in configDB's 'VLAN' table to check if the removing port is VLAN member or not. But the field is not using anymore, all the members are having separate entries in 'VLAN_MEMBER' table with the name VlanName|MemberName.

Solution:
Change to use the separate entries in 'VLAN_MEMBER' table to handle VLAN members.