JIRA-SONIC-1790: Fix the error on AS9716-32D when changing breakout mode from 4x100G to 1x100G[40G](4)
When changing breakout mode from 4x100G to 1x100G[40G](4) on as9716,
the first breakout port speed does not change before and after execution
but it uses different lane number. "4x100G" uses 2 lanes per 100G port.
"1x100G[40G](4)" uses 4 lanes per 100G port.

For the current implementation, it will not do remove the port and add
the port if it finds that the port configuration is the same before and
after executing DPB CLI.

By the constraints of the SDK and the hardware capabilities, it has to
remove all of the original ports and create the first port with 100G
using 4 hardware lanes. So we add the check against the lane number used
by the port before and after execution.